added test files

+++ src/ft_strlen.c
@@ -1,4 +1,4 @@
-int ft_strlen(char *str)
+size_t ft_strlen(char *str)
 {
     int i;
